pyproject.toml
[project]
name = "engineering-assignment-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlflow>=2.20.3",
    "pillow>=11.1.0",
    "psutil>=7.0.0",
    "pyrsmi>=0.2.0",
    "pytorch-lightning>=2.5.0.post0",
    "torch>=2.6.0",
    "torchmetrics>=1.6.2",
    "torchvision>=0.21.0",
]
